pyproject.toml
[project]
name = "crawjud"
version = "0.1.0"
description = ""
authors = [{ name = "Robotz213", email = "nicholas@robotz.dev" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "google-cloud-storage (>=3.0.0,<4.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pypdf (>=5.3.0,<6.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "clear (>=2.0.0,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "python-socketio (>=5.12.1,<6.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pywinauto (>=0.6.9,<0.7.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pymupdf (>=1.25.3,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "flask-talisman (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "deep-translator (>=1.11.4,<2.0.0)",
    "chrome-version (>=0.3.1,<0.4.0)",
    "findmyip (>=1.2.0,<2.0.1)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "pygithub (>=2.5.0,<3.0.0)",
    "openai (>=1.68.0,<2.0.1)",
    "redis (>=5.2.1,<6.0.0)",
    "redis-flask (>=0.0.2,<0.0.3)",
    "tenacity (>=9.0.0,<10.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "tornado (>=6.4.2,<7.0.0)",
    "python-dotenv-vault (>=0.6.4,<0.7.0)",
    "cryptography (<42.0.0)",
    "python-dotenv (<0.22.0)",
    "quart (>=0.20.0,<0.21.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "quart-flask-patch (>=0.3.0,<0.4.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "quart-auth (>=0.11.0,<0.12.0)",
    "aiopath (>=0.7.7,<0.8.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "inquirer (>=3.4.0,<4.0.0)",
    "termcolor (>=2.5.0,<3.0.0)",
    "pre-commit-html (>=0.1.5,<0.2.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-mysqldb (>=2.0.0,<3.0.0)",
    "reportlab (>=4.3.1,<5.0.0)",
    "babel (>=2.17.0,<3.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "flask-cloudflared (>=0.0.14,<0.0.15)",
    "gevent (>=24.11.1,<25.0.0)",
    "gevent-websocket (>=0.10.1,<0.11.0)",
    "eventlet (>=0.39.0,<0.40.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "yamllint (>=1.35.1,<2.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "quart-login @ git+https://github.com/0000matteo0000/quart-login.git",
    "quart-wtforms (>=1.0.3,<2.0.0)",
    "pynput (>=1.7.7,<2.0.0)",
    "tailer (>=0.4.1,<0.5.0)",
    "prompt-toolkit (>=3.0.50,<4.0.0)",
    "click-repl (>=0.3.0,<0.4.0)",
    "pythonnet (==3.0.5)",
    "pywebview (>=5.4,<6.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "quart-jwt-extended (>=0.1.0,<0.2.0)",
    "quart-cors (>=0.8.0,<0.9.0)",
    "billiard (>=4.2.1,<5.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
]
[tool.poetry.group.dev.dependencies]
ruff = "^0.11.1"
pre-commit = "^4.1.0"
debugpy = "^1.8.12"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
yamllint = "^1.35.1"
pytest-asyncio = "^0.25.3"

[tool.poetry.scripts]
crawjud = "crawjud.core:main_server"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B603", "B404", "B607", "B605", "B104"]
